requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
joblib==1.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.7.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
